refactor: remove unused images and update imports for consistency in ladder and FBD components feat: create generic block variant type feat: fix types of fbd
