+++ src/renderer/components/_atoms/graphical-editor/ladder/coil.tsx
@@ -14,9 +14,9 @@ import type { ReactNode } from 'react'
 import { useEffect, useRef, useState } from 'react'
 
 import { HighlightedTextArea } from '../../highlighted-textarea'
-import { getLadderPouVariablesRungNodeAndEdges } from '../utils'
 import { buildHandle, CustomHandle } from './handle'
-import type { BasicNodeData, BuilderBasicProps } from './utils/types'
+import type { BasicNodeData, BuilderBasicProps } from './utils'
+import { getLadderPouVariablesRungNodeAndEdges } from './utils'
 import { VariablesBlockAutoComplete } from './variables-block-autocomplete'
 
 export type CoilNode = Node<
