+++ src/renderer/components/_features/[workspace]/editor/graphical/elements/ladder/block/library.tsx
