+++ src/renderer/components/_features/[workspace]/editor/graphical/elements/ladder/mockImages/Group112.png
