+++ src/renderer/components/_features/[workspace]/editor/graphical/elements/ladder/mockImages/image1.png
