+++ src/renderer/store/slices/project/validation/variables.ts
@@ -1,4 +1,5 @@
 import { PLCGlobalVariable, PLCStructureVariable, PLCVariable } from '@root/types/PLC/open-plc'
+import type { PLCVariable as PLCVariableUnit } from '@root/types/PLC/units/variable'
 
 import { ProjectResponse } from '../types'
 
@@ -188,6 +189,38 @@ const checkVariableName = (variables: PLCVariable[], variableName: string) => {
     variableName.length - extractNumberAtEnd(variableName).length,
   )
   const filteredVariables = variables.filter((variable: PLCVariable) =>
+    variable.name.toLowerCase().includes(variableNameWithoutNumber.toLowerCase()),
+  )
+
+  // If there is a variable with the same name, sort the variables by the number at the end and get the biggest number
+  const sortedVariables = filteredVariables.sort((a, b) => {
+    const numberA = extractNumberAtEnd(a.name).number
+    const numberB = extractNumberAtEnd(b.name).number
+    if (numberA && numberB) {
+      return numberA - numberB
+    }
+    return 0
+  })
+
+  // Get the biggest number at the end of the variable name
+  // If there is no number at the end of the variable name, return -1 (because the number at the end of the variable name is 0)
+  const biggestVariable =
+    sortedVariables.length > 0 ? extractNumberAtEnd(sortedVariables[sortedVariables.length - 1].name) : { number: -1 }
+
+  return {
+    ok: filteredVariables.length > 0,
+    name: variableNameWithoutNumber,
+    number: biggestVariable.number + 1,
+  }
+}
+
+const checkVariableNameUnit = (variables: PLCVariableUnit[], variableName: string) => {
+  // Check if there is a variable with the same name when removing the number at the end
+  const variableNameWithoutNumber = variableName.substring(
+    0,
+    variableName.length - extractNumberAtEnd(variableName).length,
+  )
+  const filteredVariables = variables.filter((variable: PLCVariableUnit) =>
     variable.name.toLowerCase().includes(variableNameWithoutNumber.toLowerCase()),
   )
 
@@ -454,6 +487,7 @@ const updateGlobalVariableValidation = (
 export {
   arrayValidation,
   checkVariableName,
+  checkVariableNameUnit,
   createGlobalVariableValidation,
   createVariableValidation,
   enumeratedValidation,
